refactor: select bank account user + fixed translation issues
